Extracting serviceGroupApiMethods to avoid duplication of code
