+++ x-pack/test/apm_api_integration/tests/service_groups/save_service_group.spec.ts
@@ -8,85 +8,44 @@ import expect from '@kbn/expect';
 import { ApmApiError } from '../../common/apm_api_supertest';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { expectToReject } from '../../common/utils/expect_to_reject';
+import {
+  createServiceGroupApi,
+  deleteAllServiceGroups,
+  getServiceGroupsApi,
+} from './service_groups_api_methods';
 
 export default function ApiTest({ getService }: FtrProviderContext) {
   const registry = getService('registry');
   const apmApiClient = getService('apmApiClient');
 
-  async function createServiceGroupApi({
-    serviceGroupId,
-    groupName,
-    kuery,
-    description,
-    color,
-  }: {
-    serviceGroupId?: string;
-    groupName: string;
-    kuery: string;
-    description?: string;
-    color?: string;
-  }) {
-    const response = await apmApiClient.writeUser({
-      endpoint: 'POST /internal/apm/service-group',
-      params: {
-        query: {
-          serviceGroupId,
-        },
-        body: {
-          groupName,
-          kuery,
-          description,
-          color,
-        },
-      },
-    });
-    return response;
-  }
-
-  async function getServiceGroupsApi() {
-    return apmApiClient.writeUser({
-      endpoint: 'GET /internal/apm/service-groups',
-    });
-  }
-
-  async function deleteAllServiceGroups() {
-    return await getServiceGroupsApi().then((response) => {
-      const promises = response.body.serviceGroups.map((item) => {
-        if (item.id) {
-          return apmApiClient.writeUser({
-            endpoint: 'DELETE /internal/apm/service-group',
-            params: { query: { serviceGroupId: item.id } },
-          });
-        }
-      });
-      return Promise.all(promises);
-    });
-  }
-
   registry.when('Service group create', { config: 'basic', archives: [] }, () => {
-    afterEach(deleteAllServiceGroups);
+    afterEach(async () => {
+      await deleteAllServiceGroups(apmApiClient);
+    });
 
     it('creates a new service group', async () => {
       const serviceGroup = {
         groupName: 'synthbeans',
         kuery: 'service.name: synth*',
       };
-      const createResponse = await createServiceGroupApi(serviceGroup);
+      const createResponse = await createServiceGroupApi({ apmApiClient, ...serviceGroup });
       expect(createResponse.status).to.be(200);
       expect(createResponse.body).to.have.property('id');
       expect(createResponse.body).to.have.property('groupName', serviceGroup.groupName);
       expect(createResponse.body).to.have.property('kuery', serviceGroup.kuery);
       expect(createResponse.body).to.have.property('updatedAt');
-      const serviceGroupsResponse = await getServiceGroupsApi();
+      const serviceGroupsResponse = await getServiceGroupsApi(apmApiClient);
       expect(serviceGroupsResponse.body.serviceGroups.length).to.be(1);
     });
 
     it('handles invalid fields with error response', async () => {
-      const err = await expectToReject<ApmApiError>(() =>
-        createServiceGroupApi({
-          groupName: 'synthbeans',
-          kuery: 'service.name: synth* or transaction.type: request',
-        })
+      const err = await expectToReject<ApmApiError>(
+        async () =>
+          await createServiceGroupApi({
+            apmApiClient,
+            groupName: 'synthbeans',
+            kuery: 'service.name: synth* or transaction.type: request',
+          })
       );
 
       expect(err.res.status).to.be(400);
